restructure version/project helper specs
