+++ spec/helpers/projects_helper_spec.rb
@@ -32,74 +32,6 @@ describe ProjectsHelper, type: :helper do
   include ApplicationHelper
   include ProjectsHelper
 
-  before do
-    User.delete_all
-    Version.delete_all
-    Project.delete_all
-
-    set_language_if_valid('en')
-    User.current = nil
-  end
-
-  let(:test_project) { FactoryBot.create :valid_project }
-
-  describe 'a version' do
-    let(:version) { FactoryBot.create :version, project: test_project }
-
-    it 'can be formatted' do
-      expect(format_version_name(version)).to eq("#{test_project.name} - #{version.name}")
-    end
-
-    it 'can be formatted within a project' do
-      @project = test_project
-      expect(format_version_name(version)).to eq(version.name)
-    end
-
-    it 'does not create a link, without permission' do
-      expect(link_to_version(version)).to eq("#{test_project.name} - #{version.name}")
-    end
-
-    describe 'with a valid user' do
-      let(:user) { FactoryBot.create :user, member_in_project: test_project }
-      before do login_as(user) end
-
-      it 'generates a link' do
-        expect(link_to_version(version)).to eq("<a href=\"/versions/#{version.id}\">#{test_project.name} - #{version.name}</a>")
-      end
-
-      it 'generates a link within a project' do
-        @project = test_project
-        expect(link_to_version(version)).to eq("<a href=\"/versions/#{version.id}\">#{version.name}</a>")
-      end
-    end
-
-    describe 'when generating options-tags' do
-      it 'generates nothing without a version' do
-        expect(version_options_for_select([])).to be_empty
-      end
-
-      it 'generates an option tag' do
-        expect(version_options_for_select([], version)).to eq("<option selected=\"selected\" value=\"#{version.id}\">#{version.name}</option>")
-      end
-    end
-  end
-
-  describe 'a system version' do
-    let(:version) { FactoryBot.create :version, project: test_project, sharing: 'system' }
-
-    it 'can be formatted' do
-      expect(format_version_name(version)).to eq("#{test_project.name} - #{version.name}")
-    end
-  end
-
-  describe 'an invalid version' do
-    let(:version) { Object }
-
-    it 'does not generate a link' do
-      expect(link_to_version(Object)).to be_empty
-    end
-  end
-
   describe '#projects_with_level' do
     let(:root) do
       stub_descendant_of
